refactor: refactor test utils
