+++ src/common/meta/src/ddl/tests/create_table.rs
@@ -17,42 +17,24 @@ use std::collections::HashMap;
 use std::sync::Arc;
 
 use api::v1::meta::Partition;
-use api::v1::region::{QueryRequest, RegionRequest};
 use api::v1::{ColumnDataType, SemanticType};
-use common_error::ext::{BoxedError, ErrorExt};
+use common_error::ext::ErrorExt;
 use common_error::status_code::StatusCode;
 use common_procedure::{Context as ProcedureContext, Procedure, ProcedureId, Status};
 use common_procedure_test::MockContextProvider;
-use common_recordbatch::SendableRecordBatchStream;
-use common_telemetry::debug;
 
-use crate::datanode_manager::HandleResponse;
 use crate::ddl::create_table::CreateTableProcedure;
 use crate::ddl::test_util::columns::TestColumnDefBuilder;
 use crate::ddl::test_util::create_table::{
     build_raw_table_info_from_expr, TestCreateTableExprBuilder,
 };
-use crate::error;
-use crate::error::{Error, Result};
+use crate::ddl::test_util::datanode_handler::{
+    NaiveDatanodeHandler, RetryErrorDatanodeHandler, UnexpectedErrorDatanodeHandler,
+};
+use crate::error::Error;
 use crate::key::table_route::TableRouteValue;
-use crate::peer::Peer;
 use crate::rpc::ddl::CreateTableTask;
-use crate::test_util::{new_ddl_context, MockDatanodeHandler, MockDatanodeManager};
-
-#[async_trait::async_trait]
-impl MockDatanodeHandler for () {
-    async fn handle(&self, _peer: &Peer, _request: RegionRequest) -> Result<HandleResponse> {
-        unreachable!()
-    }
-
-    async fn handle_query(
-        &self,
-        _peer: &Peer,
-        _request: QueryRequest,
-    ) -> Result<SendableRecordBatchStream> {
-        unreachable!()
-    }
-}
+use crate::test_util::{new_ddl_context, MockDatanodeManager};
 
 fn test_create_table_task(name: &str) -> CreateTableTask {
     let create_table = TestCreateTableExprBuilder::default()
@@ -174,32 +156,6 @@ async fn test_on_prepare_with_no_partition_err() {
         .contains("The number of partitions must be greater than 0"),);
 }
 
-#[derive(Clone)]
-pub struct RetryErrorDatanodeHandler;
-
-#[async_trait::async_trait]
-impl MockDatanodeHandler for RetryErrorDatanodeHandler {
-    async fn handle(&self, peer: &Peer, request: RegionRequest) -> Result<HandleResponse> {
-        debug!("Returning retry later for request: {request:?}, peer: {peer:?}");
-        Err(Error::RetryLater {
-            source: BoxedError::new(
-                error::UnexpectedSnafu {
-                    err_msg: "retry later",
-                }
-                .build(),
-            ),
-        })
-    }
-
-    async fn handle_query(
-        &self,
-        _peer: &Peer,
-        _request: QueryRequest,
-    ) -> Result<SendableRecordBatchStream> {
-        unreachable!()
-    }
-}
-
 #[tokio::test]
 async fn test_on_datanode_create_regions_should_retry() {
     common_telemetry::init_default_ut_logging();
@@ -218,28 +174,6 @@ async fn test_on_datanode_create_regions_should_retry() {
     assert!(error.is_retry_later());
 }
 
-#[derive(Clone)]
-pub struct UnexpectedErrorDatanodeHandler;
-
-#[async_trait::async_trait]
-impl MockDatanodeHandler for UnexpectedErrorDatanodeHandler {
-    async fn handle(&self, peer: &Peer, request: RegionRequest) -> Result<HandleResponse> {
-        debug!("Returning mock error for request: {request:?}, peer: {peer:?}");
-        error::UnexpectedSnafu {
-            err_msg: "mock error",
-        }
-        .fail()
-    }
-
-    async fn handle_query(
-        &self,
-        _peer: &Peer,
-        _request: QueryRequest,
-    ) -> Result<SendableRecordBatchStream> {
-        unreachable!()
-    }
-}
-
 #[tokio::test]
 async fn test_on_datanode_create_regions_should_not_retry() {
     common_telemetry::init_default_ut_logging();
@@ -256,25 +190,6 @@ async fn test_on_datanode_create_regions_should_not_retry() {
     };
     let error = procedure.execute(&ctx).await.unwrap_err();
     assert!(!error.is_retry_later());
-}
-
-#[derive(Clone)]
-pub struct NaiveDatanodeHandler;
-
-#[async_trait::async_trait]
-impl MockDatanodeHandler for NaiveDatanodeHandler {
-    async fn handle(&self, peer: &Peer, request: RegionRequest) -> Result<HandleResponse> {
-        debug!("Returning Ok(0) for request: {request:?}, peer: {peer:?}");
-        Ok(HandleResponse::new(0))
-    }
-
-    async fn handle_query(
-        &self,
-        _peer: &Peer,
-        _request: QueryRequest,
-    ) -> Result<SendableRecordBatchStream> {
-        unreachable!()
-    }
 }
 
 #[tokio::test]
