+++ src/common/meta/src/ddl/tests/drop_database.rs
@@ -20,8 +20,8 @@ use common_procedure_test::MockContextProvider;
 use futures::TryStreamExt;
 
 use crate::ddl::drop_database::DropDatabaseProcedure;
+use crate::ddl::test_util::datanode_handler::{NaiveDatanodeHandler, RetryErrorDatanodeHandler};
 use crate::ddl::test_util::{create_logical_table, create_physical_table};
-use crate::ddl::tests::create_table::{NaiveDatanodeHandler, RetryErrorDatanodeHandler};
 use crate::key::schema_name::SchemaNameKey;
 use crate::test_util::{new_ddl_context, MockDatanodeManager};
